refactor: use typescript on Home
